blog_app_react-django/requirements.txt
asgiref==3.7.2
blinker==1.8.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
Django==4.2
djangorestframework==3.14.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
locust==2.29.1
MarkupSafe==2.1.5
msgpack==1.0.8
psutil==5.9.8
pycparser==2.22
pywin32==306
pyzmq==26.0.3
requests==2.31.0
roundrobin==0.0.4
sqlparse==0.4.4
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3
